+++ activesupport/lib/active_support/core_ext/object/inclusion.rb
@@ -16,12 +16,12 @@ class Object
   # Returns the receiver if it's included in the argument otherwise returns +nil+.
   # Argument must be any object which responds to +#include?+. Usage:
   #
-  #   params[:bucket_type].present_in %w( project calendar )
+  #   params[:bucket_type].presence_in %w( project calendar )
   #
   # This will throw an ArgumentError if the argument doesn't respond to +#include?+.
   #
   # @return [Object]
-  def present_in(another_object)
+  def presence_in(another_object)
     self.in?(another_object) ? self : nil
   end
 end