+++ activesupport/CHANGELOG.md
@@ -1,4 +1,4 @@
-*   Added `Object#present_in` to simplify value whitelisting.
+*   Added `Object#presence_in` to simplify value whitelisting.
 
     Before:
 
@@ -6,7 +6,7 @@
 
     After:
 
-        params[:bucket_type].present_in %w( project calendar )
+        params[:bucket_type].presence_in %w( project calendar )
 
     *DHH*